+++ Samples/iOS-Swift/iOS-Swift/Tools/SentryUIApplication.h
@@ -1,51 +1,10 @@
-@class UIApplication;
-@class UIScene;
-@class UIWindow;
-@class UIViewController;
-@protocol UIApplicationDelegate;
-
-typedef NS_ENUM(NSInteger, UIApplicationState);
+/**
+ * This header exposes a private API of the SDK for testing.
+ */
 
 NS_ASSUME_NONNULL_BEGIN
 
-/**
- * A helper tool to retrieve informations from the application instance.
- */
 @interface SentryUIApplication : NSObject
-
-/**
- * Returns the application state available at @c NSApplication.sharedApplication.applicationState
- */
-@property (nonatomic, readonly) UIApplicationState applicationState;
-
-/**
- * Application shared UIApplication instance.
- */
-@property (nonatomic, readonly, nullable) UIApplication *sharedApplication;
-
-/**
- * All application open windows.
- */
-@property (nonatomic, readonly, nullable) NSArray<UIWindow *> *windows;
-
-/**
- * Retrieves the application delegate for given UIApplication
- */
-- (nullable id<UIApplicationDelegate>)getApplicationDelegate:(UIApplication *)application;
-
-/**
- * Retrieves connected scenes for given UIApplication
- */
-- (NSArray<UIScene *> *)getApplicationConnectedScenes:(UIApplication *)application
-    API_AVAILABLE(ios(13.0), tvos(13.0));
-
-/**
- * Retrieve the topmost view controllers currently presented in the app.
- * In the case of a multi-screen iPad app, it will contain the most relevant
- * view controller for each screen. If the topmost view controller
- * is a split view controller, all of its sides are reported.
- */
-- (nullable NSArray<UIViewController *> *)relevantViewControllers;
 
 /**
  * Use @c [SentryUIApplication relevantViewControllers] and convert the