+++ Sources/Sentry/include/SentryUIApplication.h
@@ -44,14 +44,6 @@ NS_ASSUME_NONNULL_BEGIN
     API_AVAILABLE(ios(13.0), tvos(13.0));
 
 /**
- * Retrieve the topmost view controllers currently presented in the app.
- * In the case of a multi-screen iPad app, it will contain the most relevant
- * view controller for each screen. If the topmost view controller
- * is a split view controller, all of its sides are reported.
- */
-- (nullable NSArray<UIViewController *> *)relevantViewControllers;
-
-/**
  * Use @c [SentryUIApplication relevantViewControllers] and convert the
  * result to a string array with the class name of each view controller.
  */