+++ Samples/iOS-Swift/iOS-Swift/ViewControllers/PermissionsViewController.swift
@@ -57,8 +57,8 @@ class PermissionsViewController: UIViewController {
         locationManager.requestWhenInUseAuthorization()
     }
     
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
         SentrySDK.reportFullyDisplayed()
     }