+++ Tests/SentryTests/SentryCrash/SentryCrashInstallationTests.m
@@ -40,14 +40,7 @@ NS_ASSUME_NONNULL_BEGIN
     [SentryDependencyContainer reset];
 }
 
-- (SentryCrashTestInstallation *)getSut
-{
-    SentryCrashTestInstallation *installation =
-        [[SentryCrashTestInstallation alloc] initForTesting];
-    self.notificationCenter = [[TestNSNotificationCenterWrapper alloc] init];
-    [[SentryCrash sharedInstance] setSentryNSNotificationCenterWrapper:self.notificationCenter];
-    return installation;
-}
+#pragma mark - Tests
 
 - (void)testUninstall
 {
@@ -55,7 +48,8 @@ NS_ASSUME_NONNULL_BEGIN
 
     [installation install:@"/tmp"];
 
-    SentryCrashMonitorType monitorsAfterInstall = [SentryCrash sharedInstance].monitoring;
+    SentryCrashMonitorType monitorsAfterInstall
+        = SentryDependencyContainer.sharedInstance.crashReporter.monitoring;
 
     [installation uninstall];
 
@@ -80,8 +74,8 @@ NS_ASSUME_NONNULL_BEGIN
 
     [installation install:@"/tmp"];
 
-    SentryCrash *sentryCrash = [SentryCrash sharedInstance];
-    SentryCrashMonitorType monitorsAfterInstall = sentryCrash.monitoring;
+    SentryCrashMonitorType monitorsAfterInstall
+        = SentryDependencyContainer.sharedInstance.crashReporter.monitoring;
     CrashHandlerData *crashHandlerDataAfterInstall = [installation g_crashHandlerData];
 
     // To ensure multiple calls in a row work
@@ -104,14 +98,26 @@ NS_ASSUME_NONNULL_BEGIN
 
 #if SentryCrashCRASH_HAS_UIAPPLICATION
     XCTAssertEqual(55, self.notificationCenter.removeObserverWithNameInvocationsCount);
-#endif
+#endif // SentryCrashCRASH_HAS_UIAPPLICATION
+}
+
+#pragma mark - Private
+
+- (SentryCrashTestInstallation *)getSut
+{
+    SentryCrashTestInstallation *installation =
+        [[SentryCrashTestInstallation alloc] initForTesting];
+    self.notificationCenter = [[TestNSNotificationCenterWrapper alloc] init];
+    [SentryDependencyContainer.sharedInstance.crashReporter
+        setSentryNSNotificationCenterWrapper:self.notificationCenter];
+    return installation;
 }
 
 - (void)assertReinstalled:(SentryCrashTestInstallation *)installation
             monitorsAfterInstall:(SentryCrashMonitorType)monitorsAfterInstall
     crashHandlerDataAfterInstall:(CrashHandlerData *)crashHandlerDataAfterInstall
 {
-    SentryCrash *sentryCrash = [SentryCrash sharedInstance];
+    SentryCrash *sentryCrash = SentryDependencyContainer.sharedInstance.crashReporter;
     XCTAssertNotEqual(NULL, [installation g_crashHandlerData]);
     XCTAssertEqual(monitorsAfterInstall, sentryCrash.monitoring);
     XCTAssertEqual(monitorsAfterInstall, sentrycrashcm_getActiveMonitors());
@@ -126,7 +132,7 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)assertUninstalled:(SentryCrashTestInstallation *)installation
      monitorsAfterInstall:(SentryCrashMonitorType)monitorsAfterInstall
 {
-    SentryCrash *sentryCrash = [SentryCrash sharedInstance];
+    SentryCrash *sentryCrash = SentryDependencyContainer.sharedInstance.crashReporter;
     XCTAssertEqual(NULL, [installation g_crashHandlerData]);
     XCTAssertEqual(SentryCrashMonitorTypeNone, sentryCrash.monitoring);
     XCTAssertEqual(SentryCrashMonitorTypeNone, sentrycrashcm_getActiveMonitors());